+++ Sources/AppStats/Core/SDKInfo.swift
@@ -8,7 +8,7 @@ enum SDKInfo {
     
     /// SDK version (semver)
     static let version = "1.0.13"
-    
+
     /// SDK build number
     static let buildNumber = "13"